+++ app/src/main/java/org/simple/clinic/overdue/AppointmentModule.kt
@@ -10,6 +10,7 @@ import org.simple.clinic.home.overdue.OverdueAppointment_Old
 import org.simple.clinic.overdue.TimeToAppointment.Days
 import org.simple.clinic.overdue.TimeToAppointment.Months
 import org.simple.clinic.overdue.TimeToAppointment.Weeks
+import org.simple.clinic.remoteconfig.ConfigReader
 import org.simple.clinic.util.preference.StringPreferenceConverter
 import org.simple.clinic.util.preference.getOptional
 import retrofit2.Retrofit
@@ -135,4 +136,7 @@ class AppointmentModule {
         Weeks(12)
     )
   }
+
+  @Provides
+  fun providePendingAppointmentsConfig(configReader: ConfigReader) = PendingAppointmentsConfig.read(configReader)
 }