+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 - Fix views flashing in the overdue list
 - When overdue appointments are loaded or are loading, render appropriate views
 - Don't show pending overdue list footer if there are no pending overdue appointments
+- Add pending list size for the default state as a firebase remote config
 - [In Progress: 27 Apr 2022] Migrate `BloodPressureHistoryScreenEffectHandler` to use view effect
 - [In Progress: 03 Jun 2022] Implement `OverdueAppointmentListItemNew` adapter